+++ app/src/main/java/com/example/teacherforboss/MainActivity.kt
@@ -5,10 +5,10 @@ import android.util.Log
 import androidx.activity.OnBackPressedCallback
 import androidx.fragment.app.Fragment
 import com.example.teacherforboss.databinding.ActivityMainBinding
-import com.example.teacherforboss.presentation.ui.bosstalkmain.basic.BossTalkMainFragment
+import com.example.teacherforboss.presentation.ui.community.boss_talk.main.basic.BossTalkMainFragment
 import com.example.teacherforboss.presentation.ui.home.HomeFragment
 import com.example.teacherforboss.presentation.ui.mypage.MyPageFragment
-import com.example.teacherforboss.presentation.ui.teachertalkmain.basic.TeacherTalkMainFragment
+import com.example.teacherforboss.presentation.ui.community.teacher_talk.main.basic.TeacherTalkMainFragment
 import com.example.teacherforboss.util.base.BindingActivity
 
 class MainActivity : BindingActivity<ActivityMainBinding>(R.layout.activity_main) {
