--- app/src/main/java/com/example/teacherforboss/data/service/BossTalkService.kt
@@ -1,4 +1,0 @@
-package com.example.teacherforboss.data.service
-
-interface BossTalkService {
-}
