+++ app/src/main/java/com/example/teacherforboss/domain/model/signup/SignupResultEntity.kt
@@ -1,4 +1,4 @@
-package com.example.teacherforboss.domain.model
+package com.example.teacherforboss.domain.model.signup
 
 data class SignupResultEntity(
     val memberId:Long,
