+++ app/src/main/java/com/example/teacherforboss/domain/repository/SignupRepository.kt
@@ -1,7 +1,7 @@
 package com.example.teacherforboss.domain.repository
 
-import com.example.teacherforboss.domain.model.SignupEntity
-import com.example.teacherforboss.domain.model.SignupResultEntity
+import com.example.teacherforboss.domain.model.signup.SignupEntity
+import com.example.teacherforboss.domain.model.signup.SignupResultEntity
 import kotlinx.coroutines.flow.Flow
 
 interface SignupRepository {
