+++ app/src/main/java/com/example/teacherforboss/domain/usecase/SignupUseCase.kt
@@ -1,9 +1,8 @@
 package com.example.teacherforboss.domain.usecase
 
-import com.example.teacherforboss.domain.model.SignupEntity
-import com.example.teacherforboss.domain.model.SignupResultEntity
+import com.example.teacherforboss.domain.model.signup.SignupEntity
+import com.example.teacherforboss.domain.model.signup.SignupResultEntity
 import com.example.teacherforboss.domain.repository.SignupRepository
-import com.example.teacherforboss.domain.repository.UserRepository
 import kotlinx.coroutines.flow.Flow
 
 class SignupUseCase(
