+++ app/src/main/java/com/example/teacherforboss/presentation/ui/community/boss_talk/main/card/BossTalkMainCardViewHolder.kt
@@ -1,4 +1,4 @@
-package com.example.teacherforboss.presentation.ui.bosstalkmain.card
+package com.example.teacherforboss.presentation.ui.community.boss_talk.main.card
 
 import androidx.recyclerview.widget.RecyclerView
 import com.example.teacherforboss.databinding.ItemBossTalkCardBinding
