+++ app/src/main/java/com/example/teacherforboss/presentation/ui/community/teacher_talk/main/Category/TeacherTalkCategory.kt
@@ -1,0 +1,4 @@
+package com.example.teacherforboss.presentation.ui.community.teacher_talk.main.Category
+data class TeacherTalkCategory(
+    val category_name: String,
+)
