+++ app/src/main/java/com/example/teacherforboss/presentation/ui/community/teacher_talk/main/Category/TeacherTalkCategoryViewHolder.kt
@@ -1,4 +1,4 @@
-package com.example.teacherforboss.presentation.ui.teachertalkmain.Category
+package com.example.teacherforboss.presentation.ui.community.teacher_talk.main.Category
 
 import androidx.recyclerview.widget.RecyclerView
 import com.example.teacherforboss.databinding.ItemTeacherTalkCategoryBinding
