+++ app/src/main/java/com/example/teacherforboss/presentation/ui/community/teacher_talk/main/card/TeacherTalkCard.kt
@@ -1,4 +1,4 @@
-package com.example.teacherforboss.presentation.ui.teachertalkmain.card
+package com.example.teacherforboss.presentation.ui.community.teacher_talk.main.card
 
 data class TeacherTalkCard (
     val question: String,
