--- app/src/main/java/com/example/teacherforboss/presentation/ui/teachertalkmain/Category/TeacherTalkCategory.kt
@@ -1,4 +1,0 @@
-package com.example.teacherforboss.presentation.ui.teachertalkmain.Category
-data class TeacherTalkCategory(
-    val category_name: String,
-)
