+++ app/src/main/res/layout/fragment_boss_talk_main.xml
@@ -13,7 +13,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent">
         <!-- scroll view-->
-        <com.example.teacherforboss.presentation.ui.teachertalkmain.NewScrollView
+        <com.example.teacherforboss.presentation.ui.community.teacher_talk.main.NewScrollView
             android:id="@+id/sv_teacher_talk_main"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
@@ -187,7 +187,7 @@
 
 
             </LinearLayout>
-        </com.example.teacherforboss.presentation.ui.teachertalkmain.NewScrollView>
+        </com.example.teacherforboss.presentation.ui.community.teacher_talk.main.NewScrollView>
 
         <!--플로팅 버튼-->
         <com.google.android.material.floatingactionbutton.FloatingActionButton
