+++ app/src/main/res/layout/fragment_teacher_talk_main.xml
@@ -13,7 +13,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent">
         <!-- scroll view-->
-        <com.example.teacherforboss.presentation.ui.teachertalkmain.NewScrollView
+        <com.example.teacherforboss.presentation.ui.community.teacher_talk.main.NewScrollView
             android:id="@+id/sv_teacher_talk_main"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
@@ -211,7 +211,7 @@
 
                 </androidx.constraintlayout.widget.ConstraintLayout>
             </LinearLayout>
-        </com.example.teacherforboss.presentation.ui.teachertalkmain.NewScrollView>
+        </com.example.teacherforboss.presentation.ui.community.teacher_talk.main.NewScrollView>
 
         <!--플로팅 버튼-->
         <com.google.android.material.floatingactionbutton.FloatingActionButton
